// src/types/listing.ts
// src/types/listing.ts
export type FormData = {
    title: string
    businessOverview: string
    area: string[]
    category: string
    subCategory: string
    initialCost: string
    marginRate: string
    contractPeriod: string
    description: string
    companyName: string
    companyUrl: string
    imageUrl: string
  }
  